# swagger (2).yaml
swagger: '2.0'
info:
  description: API Documentation using swagger
  version: v2
  title: Swagger Demo
  termsOfService: 'http://google.com'
host: 'localhost:8090'
basePath: /
tags:
  - name: voucher-controller
    description: Voucher Controller
paths:
  /api/v1/createVoucher:
    post:
      tags:
        - voucher-controller
      summary: Create new voucher
      operationId: createVoucherUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: voucher
          description: voucher
          required: true
          schema:
            $ref: '#/definitions/VoucherModel'
      responses:
        '200':
          description: Sucess
          schema:
            $ref: '#/definitions/VoucherModel'
        '201':
          description: Created
        '204':
          description: Empty Contents
          schema:
            $ref: '#/definitions/VoucherModel'
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Unknown error
  /api/v1/fetchVoucher:
    get:
      tags:
        - voucher-controller
      summary: Fetch voucher
      operationId: fetchVoucherUsingGET
      consumes:
        - application/json
      produces:
        - '*/*'
      responses:
        '200':
          description: Sucess
          schema:
            type: array
            items:
              $ref: '#/definitions/VoucherModel'
        '204':
          description: Empty Contents
          schema:
            type: array
            items:
              $ref: '#/definitions/VoucherModel'
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Unknown error
  '/api/v1/fetchVoucher/{voucherNumber}':
    get:
      tags:
        - voucher-controller
      summary: Fetch voucher by id
      operationId: fetchVoucherByIdUsingGET
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: voucherNumber
          in: path
          description: voucherNumber
          required: true
          type: integer
          format: int32
      responses:
        '200':
          description: Sucess
          schema:
            $ref: '#/definitions/VoucherModel'
        '204':
          description: Empty Contents
          schema:
            $ref: '#/definitions/VoucherModel'
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
        '500':
          description: Unknown error
definitions:
  VoucherModel:
    type: object
    properties:
      requestedQuantity:
        type: string
      voucherAmount:
        type: string
      voucherNumber:
        type: integer
        format: int32
      voucherType:
        type: string